+++ lib/commands.ml
@@ -10,7 +10,9 @@ module Init = struct
   let init ?use_sudo ?(force = false) () =
     let base_dir = Lazy.force !Config.base_dir in
     let has_config = try Sys.is_directory base_dir with _ -> false in
-    if has_config && not force then die "E: Passage init failed, a previous installation was found in %s. Run passage init with --force to overwrite)" base_dir;
+    if has_config && not force then
+      die "E: Passage init failed, a previous installation was found in %s. Run passage init with --force to overwrite"
+        base_dir;
     try
       (* create private and pub key, ask for user's name *)
       let () =
@@ -35,7 +37,10 @@ What should be the name used for your recipient identity?|}
       in
       let user_name =
         match In_channel.input_line stdin with
-        | None -> die "E: EOF while reading user name."
+        | None ->
+          die
+            "E: username input was terminated by EOF before a newline was encountered. Ensure the username is \
+             newline-terminated and try again"
         | Some line ->
           let line = String.trim line in
           let buf = Buffer.create String.(length line) in